Return success code when nothing to commit

+++ entrypoint.sh
@@ -20,9 +20,13 @@ fi && \
 git init && \
 git config user.name "${GITHUB_ACTOR}" && \
 git config user.email "${GITHUB_ACTOR}@users.noreply.github.com" && \
+if [ -z "$(git status --porcelain)" ]; then
+    echo "Nothing to commit" && \
+    exit 0
+fi && \
 git add . && \
 git commit -m 'Deploy to GitHub pages' && \
-git push --force $REMOTE_REPO HEAD:$REMOTE_BRANCH && \
+git push --force $REMOTE_REPO master:$REMOTE_BRANCH && \
 rm -fr .git && \
 cd $GITHUB_WORKSPACE && \
 echo "Content of $BUILD_DIR has been deployed to GitHub Pages."
